Wrote 2 apps and ran unit tests using the unittest module.
First one called city_functions gathers the name of a city and country and an option argument population and returns a formatted string.

Second one is using a class that in its init method gathers the first name, last name and annual salary of an employee. A method is used to raise the annual salary by a default or custom amount.

Ran tests for both apps and used the setUp method for the Employee class to create only one object for all unit tests.